wip: create new spacewalk-data-fsck script
The existing script is hard to understand, changing it is harder than
rewriting it. For now this script can only list packages that miss an
entry in the 'path' column and, if called with --fix, update the 'path'
column when a package was found on the file system (rpm only).